requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyMySQL==1.1.1
Jinja2==3.1.4
Faker==25.3.0
mysql-connector-python==8.4.0
PyMySQL==1.1.1
Python==3.12.3